# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  handler.CreateGatewayRequest:
    properties:
      config: {}
      description:
        type: string
      fee_fixed:
        type: number
      fee_rate:
        type: number
      icon:
        type: string
      max_amount:
        type: number
      min_amount:
        type: number
      name:
        type: string
      sort:
        type: integer
      type:
        enum:
        - alipay
        - wechat
        - stripe
        - usdt
        - epay
        type: string
    required:
    - name
    - type
    type: object
...
  handler.CreatePaymentRequest:
    properties:
      amount:
        minimum: 1
        type: number
      gateway_id:
        type: integer
      order_id:
        type: integer
    required:
    - amount
    - gateway_id
    type: object
...
  handler.ToggleRequest:
    properties:
      enabled:
        type: boolean
    type: object
...
  handler.UpdateGatewayRequest:
    properties:
      config: {}
      description:
        type: string
      fee_fixed:
        type: number
      fee_rate:
        type: number
      icon:
        type: string
      max_amount:
        type: number
      min_amount:
        type: number
      name:
        type: string
      sort:
        type: integer
      type:
        enum:
        - alipay
        - wechat
        - stripe
        - usdt
        - epay
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@anixops.com
    name: API Support
    url: https://github.com/AnixOps/anix-control/issues
  description: |-
    AnixOps Control 统一控制面 API 文档
    支持用户管理、节点管理、订阅系统、支付网关、流量转发等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/AnixOps/anix-control
  title: AnixOps Control API
  version: 3.0.0-alpha.1
paths:
...
  /admin/payment/gateways:
    get:
      consumes:
      - application/json
      description: 管理员获取所有支付网关列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取支付网关列表
      tags:
      - 管理端-支付
    post:
      consumes:
      - application/json
      description: 管理员创建新的支付网关
      parameters:
      - description: 网关信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateGatewayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建支付网关
      tags:
      - 管理端-支付
  /admin/payment/gateways/{id}:
    delete:
      consumes:
      - application/json
      description: 管理员删除指定支付网关
      parameters:
      - description: 网关ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除支付网关
      tags:
      - 管理端-支付
    put:
      consumes:
      - application/json
      description: 管理员更新指定支付网关的信息
      parameters:
      - description: 网关ID
        in: path
        name: id
        required: true
        type: integer
      - description: 网关更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateGatewayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新支付网关
      tags:
      - 管理端-支付
  /admin/payment/gateways/{id}/toggle:
    post:
      consumes:
      - application/json
      description: 管理员启用或禁用指定支付网关
      parameters:
      - description: 网关ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ToggleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 切换网关状态
      tags:
      - 管理端-支付
  /admin/payment/records:
    get:
      consumes:
      - application/json
      description: 管理员获取支付记录列表，支持分页和状态筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 支付状态
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取支付记录列表
      tags:
      - 管理端-支付
  /admin/payment/stats:
    get:
      consumes:
      - application/json
      description: 管理员获取支付统计数据，支持日期范围筛选
      parameters:
      - description: '开始日期 (格式: 2006-01-02)'
        in: query
        name: start
        type: string
      - description: '结束日期 (格式: 2006-01-02)'
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取支付统计
      tags:
      - 管理端-支付
...
  /payment/callback/{type}:
    post:
      consumes:
      - application/json
      description: 接收第三方支付平台的回调通知
      parameters:
      - description: 支付类型 (alipay/wechat/epay等)
        in: path
        name: type
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: success
          schema:
            type: string
        "400":
          description: fail
          schema:
            type: string
      summary: 支付回调
      tags:
      - 支付回调
...
  /user/payment/channels:
    get:
      consumes:
      - application/json
      description: 用户获取可用的支付渠道列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取支付渠道列表
      tags:
      - 用户端
  /user/payment/create:
    post:
      consumes:
      - application/json
      description: 用户创建支付订单
      parameters:
      - description: 支付请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建支付订单
      tags:
      - 用户端
  /user/payment/records:
    get:
      consumes:
      - application/json
      description: 用户获取自己的支付记录列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取用户支付记录
      tags:
      - 用户端
  /user/payment/status/{trade_no}:
    get:
      consumes:
      - application/json
      description: 用户查询支付订单状态
      parameters:
      - description: 订单号
        in: path
        name: trade_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 查询支付状态
      tags:
      - 用户端
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...